Move back 'drawChart' to d3chart/chart.js
